+++ posthog/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+- feat: Add an optional shutdown override to `FeatureFlagInterface` ([#299](https://github.com/PostHog/posthog-android/pull/299))
+- feat: Add `localEvaluation` to the `PostHogApi` ([#299](https://github.com/PostHog/posthog-android/pull/299))
+- feat: Add API models for local evaluation ([#299](https://github.com/PostHog/posthog-android/pull/299))
+
 ## 4.1.0 - 2025-10-20
 
 - feat: add manual captureException ([#300](https://github.com/PostHog/posthog-android/issues/300))